FinancialAutomation/requirements.txt
PyQt5==5.15.11
psycopg2-binary==2.9.9
python-dotenv==1.0.0
